requirements.txt
# Core
Django>=6.0,<6.1

# Database
mysqlclient>=2.2

# Environment
python-dotenv>=1.0

# Server
uvicorn==0.40.0
whitenoise>=6.6

# Dev tools
watchdog>=4.0
